Code cleanup and refactor to avoid circular imports.
