+++ NewareNDA/__init__.py
@@ -1,2 +1,2 @@
 from .version import __version__
-from .NewareNDA import read, _count_changes, _generate_cycle_number
+from .NewareNDA import read
